requirements.txt
joblib==1.4.2
numpy==2.2.3
pandas==2.2.3
pyaml==25.1.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
threadpoolctl==3.5.0
tzdata==2025.1
